Simplification
- all typescript
