+++ .npmignore
@@ -3,9 +3,4 @@ src/
 .gitignore
 .npmignore
 .npmrc
-index.html
-demo.png
-index.ts
-rollup.config.js
 tsconfig.json
-webpack.config.js
