--- doc/index.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>QRcode-with-logo</title>
-    <style>
-      .code {
-        display: inline-block;
-        color: #ccc;
-        background-color: #666;
-        padding: 2px 10px;
-        border-radius: 2px;
-      }
-    </style>
-  </head>
-  <body>
-    <h2>qrcode-with-logos</h2>
-    <p>You can input <span class="code">qrcode.downloadImage('qrcode')</span> to download image on console.
-      <button onclick="qrcode.downloadImage('qrcode')">download!</button>
-    </p>
-    <p>You can input <span class="code">qrcode.getCanvas().then(canvas => {})</span> to deal with the canvas.</p>
-    <p>You can input <span class="code">qrcode.getImage().then(image => {})</span> to deal with the image.</p>
-    <canvas id="canvas"></canvas>
-    <img src="" alt="" id="image" />
-    <script src="../lib/qrcode-with-logos.min.js"></script>
-    <script>
-      window.qrcode = new QrCodeWithLogo({
-        canvas: document.getElementById("canvas"),
-        content: "https://github.com/zxpsuper",
-        width: 380,
-        //   download: true,
-        image: document.getElementById("image"),
-        logo: {
-          src: "https://avatars1.githubusercontent.com/u/28730619?s=460&v=4",
-          logoSize: 0.2
-        }
-      });
-    </script>
-
-  </body>
-</html>
