--- index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>QRcode-with-logo</title>
-  </head>
-  <body>
-    <canvas id="canvas"></canvas>
-    <img src="" alt="" id="image" />
-    <script src="./src/server.ts"></script>
-  </body>
-</html>
