+++ package.json
@@ -1,59 +1,20 @@
 {
   "name": "qrcode-with-logos",
-  "version": "1.0.5",
-  "description": "",
-  "main": "lib/qrcode-with-logos.common.js",
-  "types": "types/index.d.ts",
-  "module": "lib/qrcode-with-logos.esm.js",
-  "umd": "lib/qrcode-with-logos.min.js",
-  "jsdelivr": "lib/qrcode-with-logos.min.js",
-  "unpkg": "lib/qrcode-with-logos.min.js",
-  "className": "QrCodeWithLogo",
+  "version": "2.0.0",
+  "description": "Logo inside qr code",
+  "main": "build/index.js",
+  "typings": "build/index.d.ts",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "parcel ./index.html",
-    "build:type": "tsc --module commonjs",
-    "build": "npm run build:min && npm run build:rollup",
-    "build:rollup": "rollup -c",
-    "build:min": "webpack --config ./webpack.config.js"
-  },
-  "keywords": [
-    "qr-code",
-    "logo",
-    "image",
-    "qrCode",
-    "size"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zxpsuper/QRcode-with-logo.git"
-  },
-  "author": "suporka",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/zxpsuper/QRcode-with-logo/issues"
-  },
-  "homepage": "https://github.com/zxpsuper/QRcode-with-logo#readme",
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org"
-  },
-  "devDependencies": {
-    "parcel-bundler": "^1.12.4",
-    "rollup": "^2.33.3",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-json": "^4.0.0",
-    "rollup-plugin-node-polyfills": "^0.2.1",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript": "^1.0.1",
-    "ts-loader": "^8.0.12",
-    "tslib": "^2.0.3",
-    "typescript": "^3.5.2",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "webpack": "^4.35.0",
-    "webpack-cli": "^3.3.5"
+    "build": "rimraf -rf ./build && tsc -p . -d --declarationDir build",
+    "prepare": "npm run build"
   },
   "dependencies": {
-    "qrcode": "^1.4.4"
+    "qrcode": "^1.5.3"
+  },
+  "devDependencies": {
+    "@types/qrcode": "^1.5.5",
+    "rimraf": "^5.0.7",
+    "tslib": "^2.6.2",
+    "typescript": "^5.4.5"
   }
 }
