+++ tsconfig.json
@@ -1,22 +1,25 @@
 {
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "target": "es5",
-    "module": "es2015",
-    "lib": ["es2015", "es2016", "es2017", "dom", "dom.iterable", "scripthost"],
-    "strict": true,
-    "sourceMap": false,
-    "declaration": true,
-    "allowSyntheticDefaultImports": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "declarationDir": "types",
-    "outDir": "lib",
-    "downlevelIteration": true,
-    "typeRoots": ["node_modules/@types"],
-    "skipLibCheck": true,
-    // "emitDeclarationOnly": true, // rollup 构建时需注释
-    "resolveJsonModule": true
+  "ts-node": {
+    "transpileOnly": true,
+    "compiler": "typescript"
   },
-  "include": ["index.ts"]
+  "compilerOptions": {
+    "target": "ES2019",
+    "module": "commonjs",
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+    "noImplicitAny": true,
+    "sourceMap": false,
+    "outDir": "build",
+    "baseUrl": "./src",
+    "resolveJsonModule": true,
+    "importHelpers": true,
+    "experimentalDecorators": true,
+    "allowSyntheticDefaultImports": true,
+    "paths": {
+      "@/*": [
+        "src/*"
+      ]
+    }
+  }
 }
