--- types/index.d.ts
@@ -1,2 +1,0 @@
-import QrCodeWithLogo from './src/index';
-export default QrCodeWithLogo;
