--- types/src/draw-canvas.d.ts
@@ -1,2 +1,0 @@
-import { BaseOptions } from "./model";
-export declare const renderQrCode: ({ canvas, content, width, nodeQrCodeOptions }: BaseOptions) => any;
