--- types/src/draw-logo.d.ts
@@ -1,2 +1,0 @@
-import { BaseOptions } from "./model";
-export declare const drawLogo: ({ canvas, logo }: BaseOptions) => Promise<void>;
