--- types/src/toCanvas.d.ts
@@ -1,2 +1,0 @@
-import { BaseOptions } from "./model";
-export declare const toCanvas: (options: BaseOptions) => Promise<void>;
