--- types/src/toImage.d.ts
@@ -1,5 +1,0 @@
-import { BaseOptions } from './model';
-import QrCodeWithLogo from '.';
-export declare const toImage: (options: BaseOptions, instance: QrCodeWithLogo) => Promise<any>;
-/**save image */
-export declare const saveImage: (image: HTMLImageElement, name: string) => Promise<boolean>;
